+++ src/main/kotlin/dev/xeonkryptos/xeonrobotframeworkplugin/util/RobotNames.kt
@@ -1,0 +1,6 @@
+package dev.xeonkryptos.xeonrobotframeworkplugin.util
+
+object RobotNames {
+
+    const val INIT_DOT_ROBOT = "__init__.robot"
+}