+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - New action to create robot directories with `__init__.robot` file included
+- Mark directories with a `__init__.robot` file as a robot package in project explorer
 
 ### Fixed